Add files via upload
The code's inputs are the Full Plate Number, Date, and Time. Please make sure the input is as similar as the example. Upper and lower cases are considered and also spaces between words. Plate: (Example: GTY-896) 7 characters. Date: (Example: December 31, 2000) 3 words, 2 spaces, 1 comma, and the month has an upper case letter in the beginning. Time: (Example: 09:34 or 21:59) 5 characters, 1 ':' sign.